services/basis/admin/pyproject.toml
[tool.poetry]
name = "portfolio-admin"
version = "0.1.0"
description = "Admin service"
authors = ["iizhukov <iizhukov@icloud.com>"]
packages = [
    { include = "api" },
    { include = "core" },
    { include = "schemas" },
    { include = "services" },
    { include = "workers" },
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.4"
uvicorn = "0.30.6"
httpx = "0.27.2"
motor = "3.5.1"
pymongo = "4.7.2"
confluent-kafka = "2.5.0"
grpcio = ">=1.70.0"
protobuf = ">=6.0.0"
python-multipart = "0.0.9"

[tool.poetry.dependencies.portfolio-shared]
path = "../../shared/python"
develop = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
grpcio-tools = ">=1.70.0"
mypy-protobuf = "^3.6.0"
pytest-asyncio = "0.23.0"

[tool.poetry.scripts]
generate-grpc = "scripts.generate_grpc:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

